Renamed package to co.anode.anodevpn
