+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "29.0.3"
 
     defaultConfig {
-        applicationId "com.anode.anode"
+        applicationId 'co.anode.anodevpn'
         minSdkVersion 26
         targetSdkVersion 29
         versionCode 2
