+++ app/src/main/java/co/anode/anodevpn/MainActivity.kt
@@ -1,23 +1,15 @@
-package com.anode.anode
+package co.anode.anodevpn
 
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.text.Html
-import android.text.Spanned
-import android.text.method.LinkMovementMethod
 import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
-import android.view.View
-import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
-import androidx.core.text.HtmlCompat
-import com.google.android.material.floatingactionbutton.FloatingActionButton
-import com.google.android.material.snackbar.Snackbar
-import org.w3c.dom.Text
+import co.anode.anodevpn.R
 import java.io.*
 
 
@@ -28,7 +20,7 @@ class MainActivity : AppCompatActivity() {
     private fun initializeApp() {
         //Create files folder
         application.filesDir.mkdir()
-        val cjdrouteFile = File(AnodeUtil().CJDNS_PATH+"/"+AnodeUtil().cjdrouteConfFile)
+        val cjdrouteFile = File(AnodeUtil().CJDNS_PATH+"/"+ AnodeUtil().cjdrouteConfFile)
         if (!cjdrouteFile.exists()) {
             //Copy cjdroute
             var `in`: InputStream?
@@ -67,7 +59,7 @@ class MainActivity : AppCompatActivity() {
             }
         }
         //Create and initialize conf file
-        if (!File(application.filesDir.toString()+"/"+AnodeUtil().cjdrouteConfFile).exists()) {
+        if (!File(application.filesDir.toString()+"/"+ AnodeUtil().cjdrouteConfFile).exists()) {
             util!!.initializeCjdrouteConfFile()
         }
     }
