+++ app/src/main/java/co/anode/anodevpn/SecondFragment.kt
@@ -1,4 +1,4 @@
-package com.anode.anode
+package co.anode.anodevpn
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.NavHostFragment
+import co.anode.anodevpn.R
 
 class SecondFragment : Fragment() {
     override fun onCreateView(
