+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.anode.anode.FirstFragment"
+        android:name="co.anode.anodevpn.FirstFragment"
         android:label="@string/first_fragment_label"
         tools:layout="@layout/fragment_first">
 
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="com.anode.anode.SecondFragment"
+        android:name="co.anode.anodevpn.SecondFragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_second">
